source/request_processing_function/requirements.txt
google-cloud-logging
google-cloud-secret-manager
google-cloud-datastore
google-cloud-storage
google-auth
requests